subscriptions: update tests removing pay fees entry point
