+++ test/helpers/utils/errors.js
@@ -72,7 +72,6 @@ const SUBSCRIPTIONS_ERRORS = {
   TOKEN_DEPOSIT_FAILED: 'CS_TOKEN_DEPOSIT_FAILED',
   TOKEN_TRANSFER_FAILED: 'CS_TOKEN_TRANSFER_FAILED',
   PERIOD_DURATION_ZERO: 'CS_PERIOD_DURATION_ZERO',
-  FEE_AMOUNT_ZERO: 'CS_FEE_AMOUNT_ZERO',
   FEE_TOKEN_NOT_CONTRACT: 'CS_FEE_TOKEN_NOT_CONTRACT',
   OVERRATED_GOVERNOR_SHARE_PCT: 'CS_OVERRATED_GOVERNOR_SHARE_PCT',
   NON_PAST_PERIOD: 'CS_NON_PAST_PERIOD',
